phantasm/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
httpx>=0.27.0
playwright>=1.43.0
anthropic>=0.25.0
python-dotenv>=1.0.0
